+++ src/xen/mem.rs
@@ -1,6 +1,6 @@
 use core::ptr;
 
-pub static mut first_p2m_pfn: *const usize = ptr::null();
+pub static mut first_p2m_pfn: *const u64 = ptr::null();
 pub static mut nr_p2m_frames: u64 = 0;
 
 /// Current end of heap. Updated by `sbrk()`
